+++ NEWS.md
@@ -3,6 +3,7 @@
   * new feature: now users can specify openai-like APIs by providing url and setting
     agent name as "generic" in setupAgent() function
   * new feature: clean_code_blocks() now automatically runs when extractCode() is called
+  * new feature: sendPrompt() now can send previous message correspondence with LLM APIs via "previous.msgs" argument
   * fix: typos in setupAgent fixed
   * fix: sendPrompt fixed which was broken due to openai API changes
   * fix: extractInstallPkg can now deal with require() calls in code blocks